Updated vchan module.
Vchan can now handle multiple nodes in the same device tree. It differentiates
between the devices by using the Major number as an ID and the Minor number
as the direction of the buffer.

The file operations utilize the read and write functions to copy messages
from user space and then perform an hvc call to inform the hypervisor that
information is available.

This module will only function using seL4 as a hypervisor with the
communication-server applications running alongside the VMM.

An example device tree node looks like:

	vchan@80ffff000 {
		compatible = "vchan";
		port = <80>;
		dir = "tx";
		reg = <0x8 0x0ffff000 0x0 0x1000>;
	};

Where the port and direction match the vchan device declaration.

Change-Id: I857f0dbcb5ca7344d7e9e3c2961a6c12079c2e96